requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate@4305033f8035defad0a87cd38e5c918e78510ba5
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.43.3
certifi==2024.8.30
charset-normalizer==3.3.2
controlnet-aux==0.0.9
datasets==2.20.0
diffusers @ git+https://github.com/huggingface/diffusers@14f6464bef677e47e1ff13a12f4ddd97e7f3e973
dill==0.3.8
einops==0.8.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
optimum-quanto @ git+https://github.com/huggingface/optimum-quanto@ca0ec6beee01e08535127acbfa5042fb1b62414d
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.3
tabulate==0.9.0
tifffile==2024.8.30
timm==0.6.7
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers@e15687fffe5c9d20598a19aeab721ae0a7580f8a
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.12.0
zipp==3.20.2
